+++ src/errorReporting.test.ts
@@ -148,6 +148,7 @@ describe('errorReporting', () => {
                     error,
                     undefined,
                     undefined,
+                    undefined,
                 );
             });
 
@@ -155,7 +156,14 @@ describe('errorReporting', () => {
                 const error = createError('Error1');
                 errorlistener!({ error, capturedBy: 'foo' });
 
-                expect(analytics.errorEvent).toHaveBeenCalledWith(undefined, error.message, error, 'foo', undefined);
+                expect(analytics.errorEvent).toHaveBeenCalledWith(
+                    undefined,
+                    error.message,
+                    error,
+                    'foo',
+                    undefined,
+                    undefined,
+                );
             });
 
             it('with productArea', () => {
@@ -166,6 +174,7 @@ describe('errorReporting', () => {
                     'RovoDev',
                     error.message,
                     error,
+                    undefined,
                     undefined,
                     undefined,
                 );
@@ -181,6 +190,7 @@ describe('errorReporting', () => {
                     error,
                     undefined,
                     undefined,
+                    undefined,
                 );
             });
 
@@ -188,14 +198,28 @@ describe('errorReporting', () => {
                 const error = createError('Error1');
                 errorlistener!({ error, errorMessage: "what's this", capturedBy: 'fii' });
 
-                expect(analytics.errorEvent).toHaveBeenCalledWith(undefined, "what's this", error, 'fii', undefined);
+                expect(analytics.errorEvent).toHaveBeenCalledWith(
+                    undefined,
+                    "what's this",
+                    error,
+                    'fii',
+                    undefined,
+                    undefined,
+                );
             });
 
             it('with productArea, custom message, and capturedBy', () => {
                 const error = createError('Error1');
                 errorlistener!({ error, errorMessage: "what's this", capturedBy: 'fii', productArea: 'RovoDev' });
 
-                expect(analytics.errorEvent).toHaveBeenCalledWith('RovoDev', "what's this", error, 'fii', undefined);
+                expect(analytics.errorEvent).toHaveBeenCalledWith(
+                    'RovoDev',
+                    "what's this",
+                    error,
+                    'fii',
+                    undefined,
+                    undefined,
+                );
             });
 
             it('with a single param', () => {
@@ -209,6 +233,7 @@ describe('errorReporting', () => {
                     error,
                     undefined,
                     'single param',
+                    undefined,
                 );
             });
 
@@ -223,6 +248,7 @@ describe('errorReporting', () => {
                     error,
                     undefined,
                     JSON.stringify(params),
+                    undefined,
                 );
             });
 
@@ -235,6 +261,7 @@ describe('errorReporting', () => {
                     undefined,
                     error.message,
                     error,
+                    undefined,
                     undefined,
                     undefined,
                 );
@@ -250,6 +277,7 @@ describe('errorReporting', () => {
                     error,
                     'NodeJS.uncaughtException',
                     undefined,
+                    undefined,
                 );
             });
 
@@ -262,6 +290,7 @@ describe('errorReporting', () => {
                     error.message,
                     error,
                     'NodeJS.uncaughtExceptionMonitor',
+                    undefined,
                     undefined,
                 );
             });
@@ -276,6 +305,7 @@ describe('errorReporting', () => {
                     error,
                     'NodeJS.unhandledRejection',
                     undefined,
+                    undefined,
                 );
             });
         });
@@ -284,13 +314,27 @@ describe('errorReporting', () => {
             it('no extra params', () => {
                 errorlistener!({ error: 'Error1' as any });
 
-                expect(analytics.errorEvent).toHaveBeenCalledWith(undefined, 'Error1', undefined, undefined, undefined);
+                expect(analytics.errorEvent).toHaveBeenCalledWith(
+                    undefined,
+                    'Error1',
+                    undefined,
+                    undefined,
+                    undefined,
+                    undefined,
+                );
             });
 
             it('with capturedBy', () => {
                 errorlistener!({ error: 'Error1' as any, capturedBy: 'foo' });
 
-                expect(analytics.errorEvent).toHaveBeenCalledWith(undefined, 'Error1', undefined, 'foo', undefined);
+                expect(analytics.errorEvent).toHaveBeenCalledWith(
+                    undefined,
+                    'Error1',
+                    undefined,
+                    'foo',
+                    undefined,
+                    undefined,
+                );
             });
 
             it('with a custom message', () => {
@@ -299,6 +343,7 @@ describe('errorReporting', () => {
                 expect(analytics.errorEvent).toHaveBeenCalledWith(
                     undefined,
                     'Error reading stream buffer: Seg fault',
+                    undefined,
                     undefined,
                     undefined,
                     undefined,