+++ src/rovo-dev/client/rovoDevApiClient.ts
@@ -1,8 +1,9 @@
-import { RovoDevLogger } from '../util/rovoDevLogger';
+import { RovoDevTelemetryProvider } from '../rovoDevTelemetryProvider';
 import {
     RovoDevCancelResponse,
     RovoDevChatRequest,
     RovoDevHealthcheckResponse,
+    RovoDevSavedPromptsResponse,
     RovoDevStatusAPIResponse,
     ToolPermissionChoice,
 } from './rovoDevApiClientInterfaces';
@@ -92,7 +93,7 @@ export class RovoDevApiClient {
         } catch (err) {
             const reason = err.cause?.code || err.message || err;
             const error = new RovoDevApiError(`Failed to fetch '${restApi} API: ${reason}'`, 0, undefined);
-            RovoDevLogger.error(error, String(reason));
+            RovoDevTelemetryProvider.logError(error, String(reason));
             throw error;
         }
 
@@ -101,7 +102,7 @@ export class RovoDevApiClient {
         } else {
             const message = `Failed to fetch '${restApi} API: HTTP ${response.status}'`;
             const error = new RovoDevApiError(message, response.status, response);
-            RovoDevLogger.error(error, message);
+            RovoDevTelemetryProvider.logError(error, message);
             throw error;
         }
     }
@@ -288,5 +289,12 @@ export class RovoDevApiClient {
                   };
 
         await this.fetchApi('/accept-mcp-terms', 'POST', JSON.stringify(message));
+    }
+
+    public async getSavedPrompts(): Promise<RovoDevSavedPromptsResponse> {
+        const response = await this.fetchApi(`/v3/prompts`, 'GET');
+
+        const jsonResponse = (await response.json()) as RovoDevSavedPromptsResponse;
+        return jsonResponse;
     }
 }